+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add a new example with a simple player-controlled object using Rapier physics
 - Properly handle layers Z-order (see #4, #7 and #9)
 - Provide a callback to attach additional components to tiles / objects colliders to improve collisions detection (#7 and #9)
+- New feature flag: `user_properties`, map Tiled user properties on tiles and objects to Bevy components (#5 and #10)
 
 ## v0.3.3